+++ tests/test_engine.py
@@ -7,13 +7,14 @@ import moddb
 
 DEFAULT = "https://www.moddb.com/engines/sage-strategy-action-game-engine"
 
+
 @patch("moddb.utils.request", new=patched_request)
 class TestEngine:
     @pytest.fixture(params=[DEFAULT], autouse=True)
     def _get_object(self, request):
         with patch("moddb.utils.request", new=patched_request) as f:
             self.engine = moddb.Engine(moddb.get_page(request.param))
-    
+
     def test_get_articles(self):
         articles = self.engine.get_articles()
         self.engine.get_articles(4)