+++ tests/test_main.py
@@ -19,18 +19,20 @@ from tests import (
     test_hardware,
     test_poll,
     test_base,
-
     test_utils,
     test_client,
 )
 
 import moddb
 
-logger = logging.getLogger('moddb')
+logger = logging.getLogger("moddb")
 logger.setLevel(logging.DEBUG)
-handler = logging.FileHandler(filename='moddb.log', encoding='utf-8', mode='w')
-handler.setFormatter(logging.Formatter('%(asctime)s:%(levelname)s:%(name)s: %(message)s'))
+handler = logging.FileHandler(filename="moddb.log", encoding="utf-8", mode="w")
+handler.setFormatter(
+    logging.Formatter("%(asctime)s:%(levelname)s:%(name)s: %(message)s")
+)
 logger.addHandler(handler)
+
 
 @patch("moddb.utils.request", new=test_utils.patched_request)
 class TestMods(test_mod.TestMod):
@@ -39,12 +41,14 @@ class TestMods(test_mod.TestMod):
         with patch("moddb.utils.request", new=test_utils.patched_request) as f:
             self.mod = moddb.Mod(moddb.get_page(request.param))
 
+
 @patch("moddb.utils.request", new=test_utils.patched_request)
 class TestGames(test_game.TestGame):
     @pytest.fixture(params=test_utils.game_urls, autouse=True)
     def _get_object(self, request):
         with patch("moddb.utils.request", new=test_utils.patched_request) as f:
             self.game = moddb.Game(moddb.get_page(request.param))
+
 
 @patch("moddb.utils.request", new=test_utils.patched_request)
 class TestEngines(test_engine.TestEngine):
@@ -53,12 +57,14 @@ class TestEngines(test_engine.TestEngine):
         with patch("moddb.utils.request", new=test_utils.patched_request) as f:
             self.engine = moddb.Engine(moddb.get_page(request.param))
 
+
 @patch("moddb.utils.request", new=test_utils.patched_request)
 class TestFiles(test_file.TestFile):
     @pytest.fixture(params=test_utils.file_urls, autouse=True)
     def _get_object(self, request):
         with patch("moddb.utils.request", new=test_utils.patched_request) as f:
             self.file = moddb.File(moddb.get_page(request.param))
+
 
 @patch("moddb.utils.request", new=test_utils.patched_request)
 class TestAddons(test_addon.TestAddon):
@@ -67,19 +73,22 @@ class TestAddons(test_addon.TestAddon):
         with patch("moddb.utils.request", new=test_utils.patched_request) as f:
             self.addon = moddb.Addon(moddb.get_page(request.param))
 
+
 @patch("moddb.utils.request", new=test_utils.patched_request)
 class TestMedias(test_media.TestMedia):
     @pytest.fixture(params=test_utils.media_urls, autouse=True)
     def _get_object(self, request):
         with patch("moddb.utils.request", new=test_utils.patched_request) as f:
             self.media = moddb.Media(moddb.get_page(request.param))
-    
+
+
 @patch("moddb.utils.request", new=test_utils.patched_request)
 class TestArticles(test_article.TestArticle):
     @pytest.fixture(params=test_utils.article_urls, autouse=True)
     def _get_object(self, request):
         with patch("moddb.utils.request", new=test_utils.patched_request) as f:
             self.article = moddb.Article(moddb.get_page(request.param))
+
 
 @patch("moddb.utils.request", new=test_utils.patched_request)
 class TestGroups(test_group.TestGroup):
@@ -88,12 +97,14 @@ class TestGroups(test_group.TestGroup):
         with patch("moddb.utils.request", new=test_utils.patched_request) as f:
             self.group = moddb.Group(moddb.get_page(request.param))
 
+
 @patch("moddb.utils.request", new=test_utils.patched_request)
 class TestTeams(test_team.TestTeam):
     @pytest.fixture(params=test_utils.team_urls, autouse=True)
     def _get_object(self, request):
         with patch("moddb.utils.request", new=test_utils.patched_request) as f:
             self.team = moddb.Team(moddb.get_page(request.param))
+
 
 @patch("moddb.utils.request", new=test_utils.patched_request)
 class TestJobs(test_job.TestJob):
@@ -102,12 +113,14 @@ class TestJobs(test_job.TestJob):
         with patch("moddb.utils.request", new=test_utils.patched_request) as f:
             self.job = moddb.Job(moddb.get_page(request.param))
 
+
 @patch("moddb.utils.request", new=test_utils.patched_request)
 class TestMembers(test_member.TestMember):
     @pytest.fixture(params=test_utils.member_urls, autouse=True)
     def _get_object(self, request):
         with patch("moddb.utils.request", new=test_utils.patched_request) as f:
             self.member = moddb.Member(moddb.get_page(request.param))
+
 
 @patch("moddb.utils.request", new=test_utils.patched_request)
 class TestPlatforms(test_platform.TestPlatform):
@@ -116,12 +129,14 @@ class TestPlatforms(test_platform.TestPlatform):
         with patch("moddb.utils.request", new=test_utils.patched_request) as f:
             self.platform = moddb.Platform(moddb.get_page(request.param))
 
+
 @patch("moddb.utils.request", new=test_utils.patched_request)
 class TestHardwares(test_hardware.TestHardware):
     @pytest.fixture(params=test_utils.hardware_urls, autouse=True)
     def _get_object(self, request):
         with patch("moddb.utils.request", new=test_utils.patched_request) as f:
             self.hardware = moddb.Hardware(moddb.get_page(request.param))
+
 
 @patch("moddb.utils.request", new=test_utils.patched_request)
 class TestSoftwares(test_software.TestSoftware):
@@ -130,6 +145,7 @@ class TestSoftwares(test_software.TestSoftware):
         with patch("moddb.utils.request", new=test_utils.patched_request) as f:
             self.software = moddb.Software(moddb.get_page(request.param))
 
+
 @patch("moddb.utils.request", new=test_utils.patched_request)
 class TestPolls(test_poll.TestPoll):
     @pytest.fixture(params=test_utils.poll_urls, autouse=True)
@@ -137,18 +153,22 @@ class TestPolls(test_poll.TestPoll):
         with patch("moddb.utils.request", new=test_utils.patched_request) as f:
             self.poll = moddb.Poll(moddb.get_page(request.param))
 
+
 @patch("moddb.utils.request", new=test_utils.patched_request)
 class TestSearches(test_base.TestSearch):
     @pytest.fixture(params=test_utils.queries, autouse=True)
     def _get_object(self, request):
         with patch("moddb.utils.request", new=test_utils.patched_request) as f:
-            self.search= moddb.search(request.param[1], query=request.param[0])
+            self.search = moddb.search(request.param[1], query=request.param[0])
+
 
 class TestLogin(test_base.TestLogin):
     pass
 
+
 class TestFrontPage(test_base.TestFrontPage):
     pass
+
 
 class TestClient(test_client.TestClient):
     pass